requirements.txt
# RoverMitra Requirements for Railway Deployment
# CPU-compatible versions to avoid CUDA issues

# Core ML libraries (CPU versions)
torch>=2.0.0,<2.5.0
torchvision>=0.15.0,<0.20.0
torchaudio>=2.0.0,<2.5.0

# Transformers and NLP
transformers>=4.30.0
tokenizers>=0.13.0
sentence-transformers>=2.2.0

# Quantization (optional, CPU-compatible)
bitsandbytes>=0.41.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# HTTP requests
requests>=2.28.0

# Environment management
python-dotenv>=1.0.0

# System monitoring
psutil>=5.9.0

# Hugging Face Hub
huggingface-hub>=0.16.0

# Additional utilities
pathlib2>=2.3.0
typing-extensions>=4.0.0